ide_ssr

Module fragments

When specifying SSR rule, you generally want to map one kind of thing to the same kind of thing: path to path, expression to expression, type to type.

The problem is, while this kind is generally obvious to the human, the ide needs to determine it somehow. We do this in a stupid way – by pasting SSR rule into different contexts and checking what works.

Functions

expr 🔒

fragment 🔒

item 🔒

pat 🔒

stmt 🔒

ty 🔒