ide_ssr

Module replacing

Code for applying replacement templates for matches that have previously been found.

Structs

ReplacementRenderer 🔒

Functions

matches_to_edit 🔒

Returns a text edit that will replace each match in matches with its corresponding replacement template. Placeholders in the template will have been substituted with whatever they matched to in the original code.

matches_to_edit_at_offset 🔒

parse_as_kind 🔒

render_replace 🔒

token_is_method_call_receiver 🔒

Returns whether token is the receiver of a method call. Note, being within the receiver of a method call doesn’t count. e.g. if the token is $a, then $a.foo() will return true, while ($a + $b).foo() or x.foo($a) will return false.