ide_ssr

Module resolving

This module is responsible for resolving paths within rules.

Structs

ResolutionScope 🔒

ResolvedPath 🔒

ResolvedPattern 🔒

ResolvedRule 🔒

Resolver 🔒

UfcsCallInfo 🔒

Functions

is_self 🔒

path_contains_type_arguments 🔒

Returns whether path or any of its qualifiers contains type arguments.

pick_node_for_resolution 🔒

Returns a suitable node for resolving paths in the current scope. If we create a scope based on a statement node, then we can’t resolve local variables that were defined in the current scope (only in parent scopes). So we find another node, ideally a child of the statement where local variable resolution is permitted.