Crate ide_ssr

Structural Search Replace

Allows searching the AST for code that matches one or more patterns and then replacing that code based on a template.

Structs

Match
Information about a match that was found.

MatchDebugInfo

MatchFinder
Searches a crate for pattern matches and possibly replaces them with something else.

SsrError

SsrMatches

SsrPattern

SsrRule

Functions

ssr_from_comment
Attempts to build an SSR MatchFinder from a comment at the given file range. If successful, returns the MatchFinder and a TextRange covering comment.