ide_ssr
Struct MatchFinder

pub struct MatchFinder<'db> { /* private fields */ }

Searches a crate for pattern matches and possibly replaces them with something else.

Implementations

impl<'db> MatchFinder<'db>

pub fn in_context( db: &'db RootDatabase, lookup_context: FilePosition, restrict_ranges: Vec<FileRange>, ) -> Result<MatchFinder<'db>, SsrError>

Constructs a new instance where names will be looked up as if they appeared at lookup_context.

pub fn at_first_file( db: &'db RootDatabase, ) -> Result<MatchFinder<'db>, SsrError>

Constructs an instance using the start of the first file in db as the lookup context.

pub fn add_rule(&mut self, rule: SsrRule) -> Result<(), SsrError>

Adds a rule to be applied. The order in which rules are added matters. Earlier rules take precedence. If a node is matched by an earlier rule, then later rules won’t be permitted to match to it.

pub fn edits(&self) -> FxHashMap<FileId, TextEdit>

Finds matches for all added rules and returns edits for all found matches.

pub fn add_search_pattern( &mut self, pattern: SsrPattern, ) -> Result<(), SsrError>

Adds a search pattern. For use if you intend to only call find_matches_in_file. If you intend to do replacement, use add_rule instead.

pub fn matches(&self) -> SsrMatches

Returns matches for all added rules.

pub fn debug_where_text_equal( &self, file_id: EditionedFileId, snippet: &str, ) -> Vec<MatchDebugInfo>

Finds all nodes in file_id whose text is exactly equal to snippet and attempts to match them, while recording reasons why they don’t match. This API is useful for command line-based debugging where providing a range is difficult.